test(hf_client): move integration tests to test_utils
Move call_hf_pbr integration tests from test_hf_client.py to
test_utils.py to match new architecture where call_hf_pbr lives
in utils module.

- Move test_call_hf_pbr_success to test_utils.py
- Move test_call_hf_pbr_wraps_exception to test_utils.py
- Update mock paths from scripts.hf_client to scripts.utils
- test_hf_client.py now only tests low-level helpers

All 40 tests passing.

Related-To: #52
